backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

gitpython==3.1.40
python-multipart==0.0.6
aiofiles==23.2.1
chardet==5.2.0
python-magic==0.4.27

sentence-transformers==2.3.1
torch==2.2.0
transformers==4.37.0

qdrant-client==1.7.3
grpcio==1.60.0

llama-cpp-python==0.2.27
openai==0.28.1
huggingface-hub==0.20.3

numpy==1.26.3
httpx==0.26.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3